% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{ferx-package}
\alias{ferx}
\alias{ferx-package}
\title{ferx: Nonlinear Mixed Effects Modeling with Rust Backend}
\description{
Fast NLME estimation using FOCE/FOCEI with automatic differentiation. Powered by a Rust engine with Enzyme autodiff for exact gradients. Supports analytical PK models (1/2-compartment) and ODE-based models.
}
\author{
\strong{Maintainer}: InsightRx \email{dev@insightrx.com}

}
