% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ferx_predict}
\alias{ferx_predict}
\title{Population predictions from a NLME model}
\usage{
ferx_predict(model, data)
}
\arguments{
\item{model}{Path to a .ferx model file}

\item{data}{Path to a NONMEM-format CSV}
}
\value{
A data.frame with columns: ID, TIME, PRED
}
\description{
Computes population-level predictions (eta = 0) for all subjects.
}
\examples{
\dontrun{
preds <- ferx_predict("warfarin.ferx", "warfarin.csv")
}

}
